% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastroll.R
\name{expanding_regression}
\alias{expanding_regression}
\title{Expanding simple linear regression on time}
\usage{
expanding_regression(y, min_periods = 2L)
}
\arguments{
\item{y}{numeric vector (dependent variable).}

\item{min_periods}{minimum number of observations required (at least 2).}
}
\value{
A list with elements \code{intercept}, \code{slope} and \code{r2}.
}
\description{
OLS of \code{y} on the time index \code{0, 1, ..., i} over a
  growing window.
}
\examples{
expanding_regression(cumsum(rnorm(100)))

}
\seealso{
\code{\link{rolling_regression}}
}
