% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastroll.R
\name{expanding_sum}
\alias{expanding_sum}
\title{Expanding sum}
\usage{
expanding_sum(x, min_periods = 1L)
}
\arguments{
\item{x}{numeric vector.}

\item{min_periods}{minimum number of valid observations required.}
}
\value{
A numeric vector of the same length as \code{x}.
}
\description{
Expanding sum
}
\examples{
expanding_sum(c(1, 2, 3))

}
\seealso{
\code{\link{expanding_mean}}
}
