% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastroll.R
\name{expanding_var}
\alias{expanding_var}
\title{Expanding variance}
\usage{
expanding_var(x, min_periods = 2L, ddof = 1L)
}
\arguments{
\item{x}{numeric vector.}

\item{min_periods}{minimum number of valid observations required.}

\item{ddof}{delta degrees of freedom (\code{1L} sample, \code{0L}
population).}
}
\value{
A numeric vector of the same length as \code{x}.
}
\description{
Expanding variance
}
\examples{
expanding_var(rnorm(50))

}
\seealso{
\code{\link{expanding_std}}
}
