% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastroll-package.R
\docType{package}
\name{fastroll-package}
\alias{fastroll}
\alias{fastroll-package}
\title{fastroll: Fast Rolling Window Statistics}
\description{
Non-finite values (NA, NaN and infinities) are treated uniformly.
}
\details{
High-performance rolling window statistics backed by a shared C++17 core.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bagol1000/fastwindow}
  \item Report bugs at \url{https://github.com/bagol1000/fastwindow/issues}
}

}
\author{
\strong{Maintainer}: bagol1000 \email{01199218@pw.edu.pl}

Authors:
\itemize{
  \item bagol1000 \email{01199218@pw.edu.pl}
}

}
\keyword{internal}
