% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastroll.R
\name{get_num_threads}
\alias{get_num_threads}
\title{Get the default OpenMP thread count}
\usage{
get_num_threads()
}
\value{
An integer.
}
\description{
Returns the number of OpenMP threads used by parallel
  functions when their \code{n_threads} argument is \code{0}.
}
\examples{
get_num_threads()

}
\seealso{
\code{\link{set_num_threads}}
}
