% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastroll.R
\name{set_num_threads}
\alias{set_num_threads}
\title{Set the default OpenMP thread count}
\usage{
set_num_threads(n)
}
\arguments{
\item{n}{a positive integer.}
}
\value{
Invisibly, \code{NULL}.
}
\description{
Sets the number of OpenMP threads used by parallel functions
  when their \code{n_threads} argument is \code{0}.
}
\examples{
set_num_threads(2)
get_num_threads()

}
\seealso{
\code{\link{get_num_threads}}
}
