% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastroll-package.R
\docType{package}
\name{fastroll-package}
\alias{fastroll}
\alias{fastroll-package}
\title{fastroll: Fast Rolling Window Statistics}
\description{
High-performance rolling window statistics backed by a shared C++17 core.
}
\author{
\strong{Maintainer}: fastwindow contributors \email{01199218@pw.edu.pl}

Authors:
\itemize{
  \item fastwindow contributors \email{01199218@pw.edu.pl}
}

}
\keyword{internal}
